Add support of jobDSL remote repository
The jobDSL can be defined in an different remote repository then the
pipeline job definition itself. This enable the dynamic creation of
jobs just by run the Jenkins/Jobs build providing the remote repository
of the jobDSL files that has tp be used.

This means adding jobs can be done without need to change the seed jobs
definition.

Minor changes:
* add support for environment variable file
* fix deprecation warning about the define [pipeline triggers](https://github.com/jenkinsci/job-dsl-plugin/wiki/Migration#pipeline-job-plugin)